+++ pkg/controllers/kusciatask/handler/failed_handler.go
@@ -20,6 +20,7 @@ import (
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
+	"github.com/secretflow/kuscia/pkg/controllers/kusciatask/common"
 	"github.com/secretflow/kuscia/pkg/controllers/kusciatask/metrics"
 	kusciaapisv1alpha1 "github.com/secretflow/kuscia/pkg/crd/apis/kuscia/v1alpha1"
 	kusciaclientset "github.com/secretflow/kuscia/pkg/crd/clientset/versioned"
@@ -36,7 +37,7 @@ type FailedHandler struct {
 }
 
 // NewFailedHandler returns a FailedHandler instance.
-func NewFailedHandler(deps *Dependencies, finishedHandler *FinishedHandler) *FailedHandler {
+func NewFailedHandler(deps *common.Dependencies, finishedHandler *FinishedHandler) *FailedHandler {
 	return &FailedHandler{
 		FinishedHandler: finishedHandler,
 		kusciaClient:    deps.KusciaClient,
