+++ pkg/controllers/kusciatask/handler/finished_handler_test.go
@@ -27,6 +27,7 @@ import (
 	kubefake "k8s.io/client-go/kubernetes/fake"
 
 	"github.com/secretflow/kuscia/pkg/common"
+	ktcommon "github.com/secretflow/kuscia/pkg/controllers/kusciatask/common"
 	kusciaapisv1alpha1 "github.com/secretflow/kuscia/pkg/crd/apis/kuscia/v1alpha1"
 	kusciafake "github.com/secretflow/kuscia/pkg/crd/clientset/versioned/fake"
 )
@@ -79,7 +80,7 @@ func TestFinishedHandler_Handle(t *testing.T) {
 	podInformer.Informer().GetStore().Add(testPod2)
 	configMapInformer.Informer().GetStore().Add(testConfigMap)
 
-	deps := &Dependencies{
+	deps := &ktcommon.Dependencies{
 		KubeClient:      kubeClient,
 		KusciaClient:    kusciaClient,
 		PodsLister:      kubeInformersFactory.Core().V1().Pods().Lister(),
