+++ pkg/controllers/kusciatask/handler/running_handler_test.go
@@ -28,6 +28,7 @@ import (
 	kubefake "k8s.io/client-go/kubernetes/fake"
 
 	"github.com/secretflow/kuscia/pkg/common"
+	ktcommon "github.com/secretflow/kuscia/pkg/controllers/kusciatask/common"
 	kusciaapisv1alpha1 "github.com/secretflow/kuscia/pkg/crd/apis/kuscia/v1alpha1"
 	kusciafake "github.com/secretflow/kuscia/pkg/crd/clientset/versioned/fake"
 	kusciainformers "github.com/secretflow/kuscia/pkg/crd/informers/externalversions"
@@ -181,7 +182,7 @@ func TestRunningHandler_Handle(t *testing.T) {
 			trgInformer.Informer().GetStore().Add(tt.trg)
 
 			go kubeInformersFactory.Start(wait.NeverStop)
-			deps := Dependencies{
+			deps := ktcommon.Dependencies{
 				KubeClient:       kubeClient,
 				KusciaClient:     kusciaClient,
 				TrgLister:        trgInformer.Lister(),
