+++ pkg/controllers/kusciatask/handler/succeeded_handler_test.go
@@ -23,6 +23,7 @@ import (
 	kubeinformers "k8s.io/client-go/informers"
 	kubefake "k8s.io/client-go/kubernetes/fake"
 
+	"github.com/secretflow/kuscia/pkg/controllers/kusciatask/common"
 	kusciaapisv1alpha1 "github.com/secretflow/kuscia/pkg/crd/apis/kuscia/v1alpha1"
 	kusciafake "github.com/secretflow/kuscia/pkg/crd/clientset/versioned/fake"
 	kusciainformers "github.com/secretflow/kuscia/pkg/crd/informers/externalversions"
@@ -42,7 +43,7 @@ func TestSucceededHandler_Handle(t *testing.T) {
 	kusciaInformerFactory := kusciainformers.NewSharedInformerFactory(kusciaClient, 0)
 	go kubeInformersFactory.Start(wait.NeverStop)
 
-	deps := Dependencies{
+	deps := common.Dependencies{
 		KubeClient:      kubeClient,
 		KusciaClient:    kusciaClient,
 		TrgLister:       kusciaInformerFactory.Kuscia().V1alpha1().TaskResourceGroups().Lister(),
